// component/types/CropperOptions.ts
export type CropperOptions = {
    path: string;
    width?: number;
    height?: number;
    cropping?: boolean;
    cropperCircleOverlay?: boolean;
    cropperActiveWidgetColor?: string;
    cropperStatusBarColor?: string;
    cropperToolbarColor?: string;
    cropperToolbarWidgetColor?: string;
    cropperToolbarTitle?: string;
};